ctr: move containers command
