--- cmd/ctr/info.go
@@ -1,35 +1,0 @@
-package main
-
-import (
-	"github.com/containerd/containerd/cmd/ctr/commands"
-	"github.com/pkg/errors"
-	"github.com/urfave/cli"
-)
-
-var containerInfoCommand = cli.Command{
-	Name:      "info",
-	Usage:     "get info about a container",
-	ArgsUsage: "CONTAINER",
-	Action: func(context *cli.Context) error {
-		id := context.Args().First()
-		if id == "" {
-			return errors.New("container id must be provided")
-		}
-		client, ctx, cancel, err := commands.NewClient(context)
-		if err != nil {
-			return err
-		}
-		defer cancel()
-		container, err := client.LoadContainer(ctx, id)
-		if err != nil {
-			return err
-		}
-		info, err := container.Info(ctx)
-		if err != nil {
-			return err
-		}
-		commands.PrintAsJSON(info)
-
-		return nil
-	},
-}
